+++ example/observations_sql.py
@@ -39,7 +39,7 @@ def sql_example():
     sql = "SELECT * FROM data WHERE parameter='temperature_air_200' AND value < -7.0;"
     log.info(f"Invoking SQL query '{sql}'")
 
-    df = observations.collect_safe()
+    df = observations.all()
     df = df.dwd.lower().io.sql(sql)
 
     print(df)